Add new cuts for anti-ID W control region
We store two variables, baselepPt0 and baselepCh0, containing the
pT and charge of the leading base lepton.

In the WCR anti-ID region, we get the WCR cuts and replace the
lepton pT c ut with a cut on baselepPt0.

Also, instead of using the "channel" variable, in the anti-ID
region we look directly at n_baseel, n_basemu, and baselepCh0
to decide whether we are in the e (+/-) or mu (+/-) channels.