Update base lepton cut for r20.7 analysis; add base lepton ptvarcone vars
This commit does two things: first, it changes the way the base-lepton cut
works in the anti-ID region for the r20.7 analysis. The cut was requiring
n_siglep == 1 (and thus n_baseleps >= 1 implicitly). We don't want to
require *any* signal leptons, so we require n_siglep == 0 and n_baseleps >= 1
explicitly.

Also, in the C++, we store the ptvarcone{20, 30} and topoetcone variables
for baseline leptons, so they can be used in cuts.