Some changes to plotting code for anti-ID region
Make the MET significance plots be log-scale-- it might be nice to do
this only for the anti-ID regions (here, the data is orders of magnitude
larger than the MC, so log-scale is nice).

I also added a condition so 'wcranti' regions are displayed as "AntiID"
rather than just "WCR" on the plot.