Include script for creating MET significance template shape plots
I modified it to run over the output of HInvPlot. Currently, it only
knows how to run over the "allmjj" bin. The exact region, and variable,
can be passed to it as a command line option.

The ratios are then computed and printed to stdout.

The error is calculated using both binomial errors, and a TEfficiency
class set to use Bayesian statistics (with uniform prior), for
comparison.