Add new features to fake lepton script: run over all regions in input file
If the option '-a' is passed to the script, it loops through all wcranti_e
regions in the input ROOT file (ignoring "ep" and "em"). If the option
"-p" is passed as well, we include the "wcranti_ep" and "wcranti_em" regions
too.

The EPS template plots now also have the ratio printed on them.