Fake leptons: save jet HT, recalculate met significance with all jets
This adds a couple of new variables to the plotting code; primarily,
a jet HT histogram gets saved for some of the fake lepton studies.
A new "alljet_metsig" is also computed (and plotted) that uses jet HT
(i.e. the sum of the pt of all jets, not just the first two).