Add variable binning for MET significance; plotting improvements
This adds a new variable binned histogram for MET significance,
for the purpose of making plots of the anti-ID region.

The plotting script now has the ability to draw a vertical
line in order to highlight/segment a region of the plot via a new
command line option. It also has the ability to change the number of
legend columns via another new command line option, to switch back
to the older behavior where one column was the default.